unified predict for sparse and dense SGD.
